chore: Restructure transports directory structure to separate vendors
